% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{a2bcovid_app}
\alias{a2bcovid_app}
\title{Web app interface to a2bcovid}
\usage{
a2bcovid_app(rstudio = FALSE)devtool
}
\arguments{
\item{rstudio}{Set to \code{TRUE} to open the app in the RStudio Viewer.
If \code{FALSE} (the default), an external web browser is launched.}
}
\description{
Web app interface to a2bcovid
}
