% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{long_to_wide}
\alias{long_to_wide}
\title{Convert patient location data for an a2bcovid analysis from long to wide
format}
\usage{
long_to_wide(long_file)
}
\arguments{
\item{long_file}{A path name to a CSV file in long format.}
}
\value{
A path name to a temporary file containing the equivalent data in
  wide format. This can be read with \code{\link{read.csv}}.

 The names of the columns in the wide and long formats are both documented
 in the \code{\link{a2bcovid}} help page, argument \code{pat_loc_file}.
}
\description{
Convert patient location data for an a2bcovid analysis from long to wide
format
}
