% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot_a2bcovid}
\alias{plot_a2bcovid}
\title{Plot results of an a2bcovid analysis}
\usage{
plot_a2bcovid(
  x,
  hi_from,
  hi_to,
  hi_col = "red",
  palette = "YlOrRd",
  direction = 1
)
}
\arguments{
\item{x}{Data frame returned by \code{\link{a2bcovid}}.}

\item{hi_from}{Character string, naming a variable in the dataframe
indicating "from" individual IDs to be highlighted in the plot. If not
supplied, then no IDs will be highlighted.}

\item{hi_to}{Character string indicating "to" individual IDs to be
highlighted, similarly.}

\item{hi_col}{Colour to use to highlight individual IDs.}

\item{palette}{Colour palette, passed to
\code{\link[ggplot2]{scale_fill_brewer}}.}

\item{direction}{Direction of colours. Defaults to 1.  Change to -1 to
reverse the order of colours.}
}
\value{
A \pkg{ggplot2} plot object.
}
\description{
Plots a grid of colours indicating likelihood of transmission paths between
each pair of individuals
}
