% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{wide_to_long}
\alias{wide_to_long}
\title{Convert patient location data for an a2bcovid analysis from wide to long
format}
\usage{
wide_to_long(wide_file)
}
\arguments{
\item{wide_file}{A path name to a CSV file in wide format.}
}
\value{
A path name to a temporary file containing the equivalent data in
  long format. This can be read with \code{\link{read.csv}}.

 The names of the columns in the wide and long formats are both documented
 in the \code{\link{a2bcovid}} help page, argument \code{pat_loc_file}.
}
\description{
Convert patient location data for an a2bcovid analysis from wide to long
format
}
