% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{mainR}
\alias{mainR}
\title{Sentence of what this function does here}
\usage{
mainR(
  data_type = 2,
  pa = 97.1875,
  pb = 0.268908,
  po = 25.625,
  smu = 1.434065,
  ssigma = 0.6612,
  ucta = 2.59321520957074,
  uctb = 3.77600606639754,
  ucto = 3.11208004146092,
  uct_mean = 6.67992,
  rate = 8e-04,
  seq_noise = 0.772469,
  threshold = 0,
  threshold_ns = 0,
  max_n = 10,
  min_qual = 0.8,
  ali_file = "Seqs_editN20_manali_plus.fa",
  pat_file = "data1.csv",
  mov_file = "data2.csv",
  ward_file = "ward_movement_network_edit_anonymised_20200811_NoPII.csv",
  noseq = 0,
  calc_thresholds = FALSE,
  diagnostic = FALSE,
  hcw_location_default = 0.5714286,
  pat_location_default = 1
)
}
\arguments{
\item{data_type}{Document this argument here}

\item{pa}{Document this argument here}

\item{pb}{Document this argument here}

\item{po}{Etc...}

\item{smu}{documentme}

\item{ssigma}{documentme}

\item{ucta}{documentme}

\item{uctb}{documentme}

\item{ucto}{documentme}

\item{uct_mean}{documentme}

\item{rate}{documentme}

\item{seq_noise}{documentme}

\item{threshold}{documentme}

\item{threshold_ns}{documentme}

\item{max_n}{documentme}

\item{min_qual}{documentme}

\item{ali_file}{documentme}

\item{pat_file}{documentme}

\item{mov_file}{documentme}

\item{ward_file}{documentme}

\item{noseq}{documentme}

\item{calc_thresholds}{documentme}

\item{diagnostic}{documentme}

\item{hcw_location_default}{documentme}

\item{pat_location_default}{documentme}
}
\value{
Document the return value here
}
\description{
Longer description of what this function does here
}
\references{
Our paper
}
\author{
Our names
}
