% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot_a2bcovid}
\alias{plot_a2bcovid}
\title{Plot results of an a2bcovid analysis}
\usage{
plot_a2bcovid(
  x,
  cluster = TRUE,
  hi_from = "from_hcw",
  hi_to = "to_hcw",
  hi_col = "red",
  hi_lab = NULL,
  palette = NULL,
  continuous = FALSE,
  direction = 1
)
}
\arguments{
\item{x}{Data frame returned by \code{\link{a2bcovid}}.}

\item{cluster}{If \code{TRUE} (the default) then the individual IDs are rearranged  using
a heuristic clustering method so that apparent clusters of infections
appear contiguously in the plot.   If \code{FALSE} then
individuals are arranged in their original order in the data.}

\item{hi_from}{Character string, naming a variable in the dataframe
indicating "from" individual IDs to be highlighted in the plot. If not
supplied, then no IDs will be highlighted.}

\item{hi_to}{Character string indicating "to" individual IDs to be
highlighted, similarly.}

\item{hi_col}{Colour to use to highlight individual IDs.}

\item{hi_lab}{Legend to describe which individuals are highlighted.  By default this is "Healthcare workers".}

\item{palette}{Colour palette, passed to
\code{\link[ggplot2]{scale_fill_brewer}}.  If omitted, a default will be chosen}

\item{continuous}{If \code{TRUE} then the p-values are plotted on a
continuous colour scale.  Currently only implemented with the default
colour palette.  If \code{FALSE} (the default), then the p-values are
classified into ranges and plotted as a categorical variable.}

\item{direction}{Direction of colours in the brewer palettes. Defaults to 1.  Change to -1 to
reverse the order of colours.}
}
\value{
A \pkg{ggplot2} plot object.
}
\description{
Plots a grid of colours indicating likelihood of transmission paths between
each pair of individuals.
}
\seealso{
\code{\link{a2bcovid}}
}
